LPS-61338 Move to the right package
