+++ modules/apps/portlet-css/portlet-css-web/src/main/java/com/liferay/portlet/css/web/portlet/configuration/icon/PortletCSSPortletConfigurationIconFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.css.web.portlet.configuration;
+package com.liferay.portlet.css.web.portlet.configuration.icon;
 
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIconFactory;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
